# assets/mes/services.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "services.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/mes/api/v2/workorder": {
      "post": {
        "summary": "工單操作",
        "description": "工單操作包含以下功能:\n - 變更工單狀態(工單狀態請參考 ./workorder/workorder.proto enum Status),\n   包含以下變更方式:\n    - 跳過工單: 將 skipWorkOrders 內所有工單狀態變更為 SKIPPED. 將工單狀態從\n               PENDING -\u003e SKIPPED, 其他狀態不可執行此操作.\n    - 投料結束工單: closeWorkOrder=指定結束工單, 工單狀態必須為 ACTIVE,\n                   同時所有的首數(batch)狀態為 BATCH_PREPARING / BATCH_CLOSING / BATCH_CLOSED /\n                   BATCH_CANCELLED (首數狀態請參考 ./workorder/workorder.proto enum BatchStatus)\n                   任一種狀態, 系統將變更工單狀態為 CLOSING, 否則不可執行操作並將錯誤回饋到\n                   WorkOrderReply.error.\n                   系統一併會把全部首數狀態 = BATCH_PREPARING 更新成 BATCH_CANCELLED 代表該首數已被取消.\n      您亦可透過指定 closeMode=ACTION_FORCE 進行強制執行, 程式同樣回饋錯誤警示並變更工單狀態.\n      若工單狀態非上述允許的狀態, 則不可執行此操作, 即使指定 closeMode=ACTION_FORCE 也無法變更狀態.\n    - 收料結束工單: finishWorkOrder=指定結束工單, 不檢查首數狀態.\n    - 執行工單: 依據 startWorkOrder 執行指定工單. 將工單狀態從 PENDING -\u003e ACTIVE, 其他狀態不可執行此操作.\n - 查詢工單資料: 依據 loadWorkOrders 查詢工單資料, 詳細操作請參考 ./workorder/workorder.proto\n               message WorkOrderRequest.LoadWorkOrders. 其查詢結果為 workOrders.\n    - 查詢工單(todosWithDetails): 表示取得工單需含該作業流程所需用到配方以及機台參數.\n - 變更首數狀態: 透過指定 batchAction 變更指定工單的首數狀態.\n\n若一個請求包含上述所有功能, 系統執行順序為 [變更首數狀態] -\u003e [變更工單狀態] -\u003e [查詢工單資料].\n\n取得TODO工單回傳資訊:\n - id: 工單號\n - process: 生產作業階段\n - batches.count: 目標生產總批次\n - batches size分三種類型表示(只會出現一種), 分別用來表示 batches.count 對應關係:\n    - unitQuantity: 為 true 的話, 表示為單位數量, 如: 成型每次生產一條生胎\n    - fixedQuantity: 每次批次產出數量為固定數值, 依研發單位制定\n    - perBatchQuantities: 每次批次產出數量不同, 派工決定該批次產出數量\n - status: 工單狀態, 請參考 ./workorder/workorder.proto enum Status\n - fedBatch: 已經投入材料的批次\n - collectedBatch: 已收取成品的批次\n - recipe: details格式為[key]value的型式, 如下說明:\n    - groupedResources : 材料\n       - STEP-${n}-MATERIALS : 表示各階段所需使用的材料, ${n}為大於0整數\n          - requirements : 材料清單\n       - TOOLS : 表示作業所需使用的製工具清單\n          - requirements : 工具清單\n    - groupedParams : 機台參數\n       - STEP-${n}-CONTROL : 表示各階段機台設定參數, ${n}為大於0整數\n          - params : 參數清單\n       - STEP-${n}-MEASUREMENT : 表示各階段機台量測參數, ${n}為大於0整數\n          - params : 參數清單\n\n下列為回傳錯誤清單:\n - Error_ERROR_BAD_REQUEST\n - Error_ERROR_BATCH_NOT_FOUND (找不到工單作業批次)\n - Error_ERROR_WORKORDER_BAD_BATCH (錯誤的作業批次)\n - Error_ERROR_WORKORDER_NOT_FOUND (找不到工單)\n - Error_ERROR_WORKORDER_BAD_STATUS (工單狀態錯誤不允許進行操作)",
        "operationId": "WorkOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiWorkOrderReply"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiWorkOrderRequest"
            }
          }
        ],
        "tags": [
          "API"
        ]
      }
    }
  },
  "definitions": {
    "BatchesPerBatch": {
      "type": "object",
      "properties": {
        "quantities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v2commonsDecimal"
          }
        }
      },
      "title": "Three possible ways to specify batch sizes:\n 1. all single-unit batches : set unitQuantity=true\n 2. all same size batches   : specify the size in fixedQuantity\n 3. different size batches  : specify batch sizes in perBatchQuantities\n 4. plan quantity : specify quantity(plan), batch sizes and batchCount in PlanQuantity"
    },
    "BatchesPlanQuantity": {
      "type": "object",
      "properties": {
        "quantity": {
          "$ref": "#/definitions/v2commonsDecimal"
        },
        "batchSize": {
          "$ref": "#/definitions/v2commonsDecimal"
        }
      }
    },
...
    "ErrorCode": {
      "type": "string",
      "enum": [
        "ERROR_NONE",
        "ERROR_INTERNAL",
        "ERROR_BAD_REQUEST",
        "ERROR_USER_NOT_FOUND_OR_BAD_PASSWORD",
        "ERROR_USER_NO_PERMISSION",
        "ERROR_USER_MISSING_TOKEN",
        "ERROR_USER_UNKNOWN_TOKEN",
        "ERROR_STATION_NOT_FOUND",
        "ERROR_STATION_ON_HOLD",
        "ERROR_RESOURCE_NOT_FOUND",
        "ERROR_RESOURCE_EXPIRED",
        "ERROR_RESOURCE_ON_HOLD",
        "ERROR_RESOURCE_EXISTED",
        "ERROR_RESOURCE_CONTROL_ABOVE_EXTENDED_COUNT",
        "ERROR_WORKORDER_NOT_FOUND",
        "ERROR_WORKORDER_BAD_BATCH",
        "ERROR_WORKORDER_BAD_STATUS",
        "ERROR_WORKORDER_UNEXPECTED_RESULTS",
        "ERROR_CARRIER_NOT_FOUND",
        "ERROR_CARRIER_IN_USE",
        "ERROR_CARRIER_FULL",
        "ERROR_BATCH_NOT_FOUND",
        "ERROR_BATCH_NOT_READY",
        "ERROR_RECORD_NOT_FOUND",
        "ERROR_RECORD_EXISTED",
        "ERROR_RECIPE_NOT_FOUND",
        "ERROR_US_MISMATCH",
        "ERROR_US_SIGN_ON_FORBIDDEN",
        "ERROR_SR_RESOURCE_IN_USE",
        "ERROR_SR_SITE_NOT_FOUND",
        "ERROR_SW_MISMATCH",
        "ERROR_RW_RESOURCE_NOT_READY",
        "ERROR_RW_QUANTITY_BELOW_MIN",
        "ERROR_RW_QUANTITY_ABOVE_MAX",
        "ERROR_RW_QUANTITY_LOW_DOSAGE",
        "ERROR_RW_QUANTITY_SHORTAGE",
        "ERROR_RW_BAD_GRADE",
        "ERROR_RW_RESOURCE_UNEXPECTED",
        "ERROR_RW_RESOURCE_MISSING"
      ],
      "default": "ERROR_NONE",
      "title": "- ERROR_INTERNAL: code value \u003c 10000 reserved for plant side application, you should use gRPC error instead\nDO NOT USE IT AS MES API ERROR.\n - ERROR_STATION_ON_HOLD: e.g., maintenance.\n - ERROR_RESOURCE_EXISTED: resource id has been used and thus not valid for new resource from Collect\n - ERROR_WORKORDER_BAD_BATCH: the batch is not allowed to be operated. e.g. CLOSED batch can not be closed.\n - ERROR_WORKORDER_BAD_STATUS: work order unexpected status, e.g: load feed work order only allowed ACTIVE status, load collect work order only allowed ACTIVE or CLOSING status.\n - ERROR_WORKORDER_UNEXPECTED_RESULTS: work order unexpected results, e.g: requested todo work order for feeding but return zero-length todo reply.\n - ERROR_CARRIER_IN_USE: carrier in use and cannot be shared\n - ERROR_CARRIER_FULL: carrier can be shared but it's full\n - ERROR_BATCH_NOT_READY: not enough fed batch to collect\n - ERROR_RECORD_NOT_FOUND: not found query record.\n - ERROR_RECORD_EXISTED: record has been collected.\n - ERROR_US_MISMATCH: the user is not the station/site operator.\n - ERROR_US_SIGN_ON_FORBIDDEN: e.g., not qualified/registered to operate the station.\n - ERROR_SR_RESOURCE_IN_USE: the resource is bound to another station/site provided in details.\n - ERROR_SW_MISMATCH: the work order is not being executed the station\n - ERROR_RW_RESOURCE_NOT_READY: details: \"required_preprocess_name\", e.g., \"cutting\"\n - ERROR_RW_QUANTITY_BELOW_MIN: the used quantity is less than the minimum quantity in the recipe.\n - ERROR_RW_QUANTITY_ABOVE_MAX: the used quantity is greater than the maximum quantity in the recipe.\n - ERROR_RW_QUANTITY_LOW_DOSAGE: the used quantity is less than the dosage.\n - ERROR_RW_QUANTITY_SHORTAGE: the quantity of the stock is less than the specified quantity(e.g. used quantity, the minimum quantity in the recipe).\n - ERROR_RW_BAD_GRADE: details: \"required_resource_grade\", e.g., \"AA\"\n - ERROR_RW_RESOURCE_UNEXPECTED: the resource is not in the work order recipe\n - ERROR_RW_RESOURCE_MISSING: the resource is required but missing"
    },
...
    "ResourceRequirementSiteRequirement": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "unique": {
          "type": "boolean",
          "format": "boolean"
        },
        "any": {
          "type": "boolean",
          "format": "boolean"
        },
        "index": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "WorkOrderReplyWorkOrders": {
      "type": "object",
      "properties": {
        "past": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v2workorderInfo"
          }
        },
        "todo": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v2workorderInfo"
          }
        }
      }
    },
    "WorkOrderRequestBatchAction": {
      "type": "object",
      "properties": {
        "workOrder": {
          "type": "string"
        },
        "batch": {
          "type": "integer",
          "format": "int64"
        },
        "status": {
          "$ref": "#/definitions/workorderBatchStatus"
        },
        "note": {
          "type": "string"
        }
      },
      "title": "Manual change to batch status"
    },
    "WorkOrderRequestLoadWorkOrders": {
      "type": "object",
      "properties": {
        "past": {
          "type": "integer",
          "format": "int64"
        },
        "todo": {
          "type": "integer",
          "format": "int64"
        },
        "pastIncludesSkipped": {
          "type": "boolean",
          "format": "boolean"
        },
        "pastIncludesClosed": {
          "type": "boolean",
          "format": "boolean"
        },
        "todosWithDetails": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
...
    "apiWorkOrderReply": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/checkError"
        },
        "workOrders": {
          "$ref": "#/definitions/WorkOrderReplyWorkOrders"
        }
      }
    },
    "apiWorkOrderRequest": {
      "type": "object",
      "properties": {
        "skipWorkOrders": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "closeWorkOrder": {
          "type": "string"
        },
        "startWorkOrder": {
          "type": "string"
        },
        "loadWorkOrders": {
          "$ref": "#/definitions/WorkOrderRequestLoadWorkOrders"
        },
        "note": {
          "type": "string"
        },
        "batchAction": {
          "$ref": "#/definitions/WorkOrderRequestBatchAction"
        },
        "closeMode": {
          "$ref": "#/definitions/checkActionMode"
        },
        "finishWorkOrder": {
          "type": "string"
        }
      }
    },
    "checkActionMode": {
      "type": "string",
      "enum": [
        "ACTION_AUTO",
        "ACTION_FORCE",
        "ACTION_PRECHECK"
      ],
      "default": "ACTION_AUTO"
    },
    "checkError": {
      "type": "object",
      "properties": {
        "code": {
          "$ref": "#/definitions/ErrorCode"
        },
        "details": {
          "type": "string"
        }
      }
    },
...
    "commonsTime": {
      "type": "object",
      "properties": {
        "nano": {
          "type": "string",
          "format": "int64",
          "title": "now.UnixNano()"
        }
      }
    },
...
    "recipeDecimalParam": {
      "type": "object",
      "properties": {
        "value": {
          "$ref": "#/definitions/v2commonsDecimal"
        },
        "min": {
          "$ref": "#/definitions/v2commonsDecimal"
        },
        "max": {
          "$ref": "#/definitions/v2commonsDecimal"
        },
        "unit": {
          "type": "string"
        }
      }
    },
    "recipeFreshness": {
      "type": "object",
      "properties": {
        "resting": {
          "$ref": "#/definitions/commonsTime"
        },
        "expiry": {
          "$ref": "#/definitions/commonsTime"
        },
        "extension": {
          "$ref": "#/definitions/v2commonsDecimal"
        },
        "maxExtensionTimes": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "recipeRecipe": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "version": {
          "type": "string"
        },
        "batchSize": {
          "$ref": "#/definitions/v2commonsDecimal"
        },
        "releasedAt": {
          "$ref": "#/definitions/commonsTime"
        },
        "details": {
          "$ref": "#/definitions/recipeRecipeDetails"
        }
      }
    },
    "recipeRecipeDetails": {
      "type": "object",
      "properties": {
        "groupedResources": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/recipeResourceRequirements"
          }
        },
        "groupedParams": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/v2recipeParams"
          }
        }
      }
    },
    "recipeResourceRequirement": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "quantity": {
          "$ref": "#/definitions/recipeDecimalParam"
        },
        "grade": {
          "type": "string"
        },
        "freshness": {
          "$ref": "#/definitions/recipeFreshness"
        },
        "site": {
          "$ref": "#/definitions/ResourceRequirementSiteRequirement"
        }
      }
    },
    "recipeResourceRequirements": {
      "type": "object",
      "properties": {
        "requirements": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/recipeResourceRequirement"
          }
        }
      }
    },
...
    "v2commonsDecimal": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "format": "int64"
        },
        "exp": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "v2recipeParams": {
      "type": "object",
      "properties": {
        "params": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v2recipeParamsParam"
          }
        }
      }
    },
    "v2recipeParamsParam": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "decimalValue": {
          "$ref": "#/definitions/recipeDecimalParam"
        },
        "stringValue": {
          "type": "string"
        }
      }
    },
...
    "v2workorderInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "process": {
          "type": "string"
        },
        "batches": {
          "$ref": "#/definitions/workorderBatches"
        },
        "recipe": {
          "$ref": "#/definitions/recipeRecipe"
        },
        "status": {
          "$ref": "#/definitions/workorderStatus"
        },
        "fedBatch": {
          "type": "integer",
          "format": "int32"
        },
        "collectedBatch": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "workorderBatchStatus": {
      "type": "string",
      "enum": [
        "BATCH_PREPARING",
        "BATCH_STARTED",
        "BATCH_CLOSING",
        "BATCH_CLOSED",
        "BATCH_CANCELLED"
      ],
      "default": "BATCH_PREPARING"
    },
    "workorderBatches": {
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int32"
        },
        "unitQuantity": {
          "type": "boolean",
          "format": "boolean"
        },
        "fixedQuantity": {
          "$ref": "#/definitions/v2commonsDecimal"
        },
        "perBatchQuantities": {
          "$ref": "#/definitions/BatchesPerBatch"
        },
        "planQuantity": {
          "$ref": "#/definitions/BatchesPlanQuantity"
        }
      }
    },
    "workorderStatus": {
      "type": "string",
      "enum": [
        "PENDING",
        "ACTIVE",
        "CLOSING",
        "CLOSED",
        "SKIPPED"
      ],
      "default": "PENDING"
    }
  }
}